+++ tests/alley/wp/alleyvate/features/test-login-nonce.php
@@ -12,6 +12,8 @@
  * @phpcs:disable WordPress.WP.GlobalVariablesOverride.Prohibited, Generic.CodeAnalysis.EmptyStatement.DetectedCatch, WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedHooknameFound
  */
 
+declare( strict_types=1 );
+
 namespace Alley\WP\Alleyvate\Features;
 
 use Mantle\Testing\Concerns\Refresh_Database;