+++ CHANGELOG.md
@@ -9,6 +9,13 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 * `prevent_framing`: Added a feature to prevent framing of the site via the
   `X-Frame-Options` header.
 
+### Changed
+
+* Unit tests: misc changes and fixes.
+* Unit tests: the `$feature` property uses the main feature class for better IDE intelephense support.
+* Unit tests: all test cases use `declare( strict_types=1 );`.
+* Unit tests: added test to confirm the attachment rewrite rules are removed
+
 ## 2.3.1
 
 ### Changed